+++ comicbox/cli.py
@@ -1,20 +1,34 @@
 """Cli for comicbox."""
+import sys
 from argparse import Action, ArgumentParser, Namespace, RawDescriptionHelpFormatter
 from collections.abc import Sequence
 
 from comicbox.config import get_config
+from comicbox.exceptions import UnsupportedArchiveTypeError
 from comicbox.metadata.comet import CoMet
 from comicbox.metadata.comicbookinfo import ComicBookInfo
 from comicbox.metadata.comicinfoxml import ComicInfoXml
 from comicbox.metadata.filename import FilenameMetadata
+from comicbox.metadata.pdf import PDFParser
 from comicbox.run import Runner
+
+WRITE_KEY_MAPS = {
+    "comet": CoMet,
+    "comicbookinfo": ComicBookInfo,
+    "comicinfoxml": ComicInfoXml,
+    "filename": FilenameMetadata,
+    "pdf": PDFParser,
+}
+
+READ_KEY_MAPS = {**WRITE_KEY_MAPS, "filename": FilenameMetadata}
+HANDLED_EXCEPTIONS = (UnsupportedArchiveTypeError,)
 
 
 class KeyValueDictAction(Action):
-    """Parse comma deliminted key value pairs key value."""
+    """Parse comma delimited key value pairs key value."""
 
     def __call__(self, _parser, namespace, values, _option_string=None):
-        """Parse comma deliminated key value pairs."""
+        """Parse comma delimited key value pairs."""
         if not values:
             return
         key, values_list = values.split("=")
@@ -31,7 +45,7 @@ class CSVAction(Action):
     """Parse comma deliminated sequences."""
 
     def __call__(self, _parser, namespace, values, _option_string=None):
-        """Parse comma deliminated sequences."""
+        """Parse comma delimited sequences."""
         if isinstance(values, str):
             values_array = values.split(",")
         elif isinstance(values, Sequence):
@@ -41,34 +55,47 @@ class CSVAction(Action):
         setattr(namespace, self.dest, values_array)
 
 
-WRITE_KEY_MAPS = {
-    "comet": CoMet,
-    "comicbookinfo": ComicBookInfo,
-    "comicinfoxml": ComicInfoXml,
-    "filename": FilenameMetadata,
-}
+class PageRangeAction(Action):
+    """Parse page range."""
 
-READ_KEY_MAPS = {**WRITE_KEY_MAPS, "filename": FilenameMetadata}
+    def __call__(self, _parser, namespace, values, _option_string=None):
+        """Parse page range delimited by :."""
+        values_array = values.split(":")
+        if not values_array:
+            return
+
+        index_from = int(values_array[0]) if len(values_array[0]) else None
+
+        if len(values_array) == 1:
+            index_to = index_from
+        elif len(values_array[1]):
+            index_to = int(values_array[1])
+        else:
+            index_to = None
+
+        if index_from is not None:
+            namespace.index_from = index_from
+        if index_to is not None:
+            namespace.index_to = index_to
 
 
 def map_keys(config, prefix, list_key, maps, value):
     """Map keyed values to config booleans."""
     key_list = getattr(config, list_key)
-    if not key_list:
-        return
-    for key in key_list:
-        lower_key = key.lower()
-        for attr_suffix, parser_class in maps.items():
-            if lower_key in parser_class.CONFIG_KEYS:
-                attr = f"{prefix}_{attr_suffix}"
-                setattr(config, attr, value)
+    if key_list:
+        for key in key_list:
+            lower_key = key.lower()
+            for attr_suffix, parser_class in maps.items():
+                if lower_key in parser_class.CONFIG_KEYS:
+                    attr = f"{prefix}_{attr_suffix}"
+                    setattr(config, attr, value)
     delattr(config, list_key)
 
 
 def process_keys(config):
     """CLI config post processing."""
     map_keys(config, "read", "ignore_read", READ_KEY_MAPS, False)
-    map_keys(config, "write", "write", READ_KEY_MAPS, True)
+    map_keys(config, "write", "write", WRITE_KEY_MAPS, True)
 
 
 def get_args(params=None) -> Namespace:
@@ -80,6 +107,7 @@ def get_args(params=None) -> Namespace:
         f"    Comic Book Info: {', '.join(sorted(ComicBookInfo.CONFIG_KEYS))}\n"
         f"    CoMet: {', '.join(sorted(CoMet.CONFIG_KEYS))}\n"
         f"    Filename: {', '.join(sorted(FilenameMetadata.CONFIG_KEYS))}\n"
+        f"    PDF: {', '.join(sorted(PDFParser.CONFIG_KEYS))}\n"
         "-w & -R can take comma separated lists of these keys."
     )
     parser = ArgumentParser(
@@ -94,12 +122,6 @@ def get_args(params=None) -> Namespace:
         action=CSVAction,
         dest="ignore_read",
         help="Ignore reading metadata formats. List of format keys.",
-    )
-    parser.add_argument(
-        "-M",
-        "--no-metadata",
-        action="store_false",
-        help="Do not read any comic metadata.",
     )
     parser.add_argument(
         "-d",
@@ -150,11 +172,17 @@ def get_args(params=None) -> Namespace:
         help="Print archive file type",
     )
     parser.add_argument(
-        "-f",
-        "--from",
-        dest="index_from",
+        "-n",
+        "--index",
+        dest="index",
         type=int,
-        help="Extract pages from the specified index forward.",
+        help="Extract a single page by zero based index.",
+    )
+    parser.add_argument(
+        "-a",
+        "--pages",
+        action=PageRangeAction,
+        help="Extract a single page or : delimited range of pages by zero based index.",
     )
     parser.add_argument(
         "-c", "--covers", action="store_true", help="Extract cover pages."
@@ -217,4 +245,8 @@ def main(params=None):
     config = get_config(args)
 
     runner = Runner(config)
-    runner.run()
+    try:
+        runner.run()
+    except HANDLED_EXCEPTIONS as exc:
+        print(exc)  # noqa: T201
+        sys.exit(1)