+++ comicbox/comic_archive.py
@@ -34,6 +34,8 @@ from comicbox.metadata.comic_xml import ComicXml
 from comicbox.metadata.comicbookinfo import ComicBookInfo
 from comicbox.metadata.comicinfoxml import ComicInfoXml
 from comicbox.metadata.filename import FilenameMetadata
+from comicbox.metadata.pdf import PDFParser
+from comicbox.pdf_file import PDFFile
 
 RECOMPRESS_SUFFIX = ".comicbox_tmp_zip"
 CBZ_SUFFIX = ".cbz"
@@ -64,7 +66,9 @@ class ComicArchive:
     _PAGES_KEYS = frozenset(frozenset(("pages",)) | _PAGE_KEYS)
     _RAW_CBI_KEY = "ComicBookInfo Archive Comment"
     _RAW_FILENAME_KEY = "Filename"
+    _RAW_PDF_KEY = "PDF"
     _MODE_EXECUTABLE = stat.S_IXUSR ^ stat.S_IXGRP ^ stat.S_IXOTH
+    _COMMENT_ARCHIVE_TYPES = (ZipFile, RarFile)
 
     def __init__(
         self,
@@ -85,8 +89,10 @@ class ComicArchive:
             config = get_config()
         self._config: AttrDict = config
         self._set_archive_cls()
-        self._archive: Union[ZipFile, RarFile, TarFile, None] = None
-        self._metadata: ComicBaseMetadata = ComicBaseMetadata(metadata=metadata)
+        self._archive: Union[ZipFile, RarFile, TarFile, PDFFile, None] = None
+        self._metadata: ComicBaseMetadata = ComicBaseMetadata(
+            path=self._path, metadata=metadata
+        )
         self._raw: dict = {}
 
     def __enter__(self):
@@ -125,6 +131,10 @@ class ComicArchive:
         elif is_tarfile(self._path):
             self._archive_cls = tarfile_open
             self._file_type = "CBT"
+        elif PDFFile.is_pdffile(self._path):
+            PDFFile.check_import()
+            self._archive_cls = PDFFile
+            self._file_type = "PDF"
         else:
             reason = f"Unsupported archive type: {self._path}"
             raise UnsupportedArchiveTypeError(reason)
@@ -159,7 +169,7 @@ class ComicArchive:
         infolist = sorted(infolist, key=lambda i: getattr(i, fn_attr))
         return infolist, fn_attr
 
-    def _archive_readfile(self, filename) -> bytes:
+    def _archive_readfile(self, filename, to_pixmap=False) -> bytes:
         """Read an archive file to memory."""
         archive = self._get_archive()
         data = b""
@@ -167,6 +177,8 @@ class ComicArchive:
             file_obj = archive.extractfile(filename)
             if file_obj:
                 data = file_obj.read()
+        elif to_pixmap and isinstance(archive, PDFFile):
+            data = archive.read(filename, True)
         else:
             data = archive.read(filename)
         return data
@@ -202,7 +214,7 @@ class ComicArchive:
         result_parsers = self._get_raw_files_metadata()
         files_md = {}
         for parser_class, data in result_parsers.items():
-            parser = parser_class(string=data)
+            parser = parser_class(path=self._path, string=data)
             files_md[parser_class] = parser.metadata
         return files_md
 
@@ -213,7 +225,12 @@ class ComicArchive:
             and self._config.read_comicbookinfo
             and not self._raw.get(self._RAW_CBI_KEY)
         ):
-            comment = self._get_archive().comment  # type: ignore
+            archive = self._get_archive()
+            if isinstance(archive, self._COMMENT_ARCHIVE_TYPES):
+                comment = archive.comment
+            else:
+                comment = None
+
             if isinstance(comment, bytes):
                 comment = comment.decode(errors="replace")
             if comment:
@@ -225,22 +242,42 @@ class ComicArchive:
         data = self._get_raw_archive_comment()
         if not data:
             return {}
-        parser = ComicBookInfo(string=data, path=self._path)
+        parser = ComicBookInfo(path=self._path, string=data)
+        return parser.metadata
+
+    def _get_raw_pdf_metadata(self):
+        if self._RAW_PDF_KEY not in self._raw:
+            if self._archive_cls != PDFFile:
+                return None
+            archive = self._get_archive()
+            if not isinstance(archive, PDFFile):
+                return None
+            raw = archive.get_metadata()
+            self._raw[self._RAW_PDF_KEY] = raw
+        return self._raw[self._RAW_PDF_KEY]
+
+    def _parse_pdf_metadata(self):
+        pdf_md = self._get_raw_pdf_metadata()
+        parser = PDFParser(path=self._path, native_dict=pdf_md)
         return parser.metadata
 
     def _get_raw_filename(self):
         """Get the filename form the path."""
-        if self._config.read_filename and not self._raw.get(self._RAW_FILENAME_KEY):
+        if not self._config.read_filename:
+            return None
+        if not self._raw.get(self._RAW_FILENAME_KEY):
             data = self._path.name
             self._raw[self._RAW_FILENAME_KEY] = data
         return self._raw.get(self._RAW_FILENAME_KEY)
 
     def _parse_metadata_filename(self):
         """Parse metadata from the filename."""
+        if not self._config.read_filename:
+            return None
         data = self._get_raw_filename()
         if data is None:
             return {}
-        parser = FilenameMetadata(path=data)
+        parser = FilenameMetadata(path=data, metadata_path=data)
         return parser.metadata
 
     def _ensure_page_metadata(self):
@@ -250,7 +287,8 @@ class ComicArchive:
             or self._metadata.get_num_pages() is None
         ):
             namelist = self._archive_namelist()
-            self._metadata.set_page_metadata(namelist)
+            is_pdf = self._archive_cls == PDFFile
+            self._metadata.set_page_metadata(namelist, is_pdf)
 
     def _normalize_metadata(self):
         """Map familiar keys to native format keys."""
@@ -267,22 +305,31 @@ class ComicArchive:
             normal_md[new_key] = value
         return normal_md
 
+    def add_metadata(self, metadata):
+        """Add metadata to the existing metadata."""
+        md_list = [metadata]
+        self._metadata.synthesize_metadata(md_list)
+
     def _parse_metadata(self):
         """Parse all enabled metadata."""
         md_list = []
         filename_md = self._parse_metadata_filename()
         if filename_md:
             md_list += [filename_md]
-        files_md = self._parse_files_metadata()
-        comet_md = files_md.get(CoMet)
-        if comet_md:
-            md_list += [comet_md]
-        cbi_md = self._parse_metadata_comments()
-        if cbi_md:
-            md_list += [cbi_md]
-        cix_md = files_md.get(ComicInfoXml)
-        if cix_md:
-            md_list += [cix_md]
+        pdf_md = self._parse_pdf_metadata()
+        if pdf_md:
+            md_list += [pdf_md]
+        if self._archive_cls != PDFFile:
+            files_md = self._parse_files_metadata()
+            comet_md = files_md.get(CoMet)
+            if comet_md:
+                md_list += [comet_md]
+            cbi_md = self._parse_metadata_comments()
+            if cbi_md:
+                md_list += [cbi_md]
+            cix_md = files_md.get(ComicInfoXml)
+            if cix_md:
+                md_list += [cix_md]
         if self._config.metadata:
             normal_md = self._normalize_metadata()
             md_list += [normal_md]
@@ -296,22 +343,26 @@ class ComicArchive:
         self._get_raw_filename()
         self._get_raw_archive_comment()
         self._get_raw_files_metadata()
+        self._get_raw_pdf_metadata()
 
     def _write_cbi_comment(self, parser):
         """Write a cbi comment to an archive."""
-        if self._archive_cls == tarfile_open:
-            reason = "Cannot write ComicBookInfo comments to cbt tarfile."
-            raise ValueError(reason)
+        if self._archive_cls not in self._COMMENT_ARCHIVE_TYPES:
+            reason = "Cannot write ComicBookInfo comments to this file."
+            raise TypeError(reason)
         self.close()
         with self._archive_cls(self._path, "a") as append_archive:
+            if not isinstance(append_archive, self._COMMENT_ARCHIVE_TYPES):
+                reason = "Cannot write ComicBookInfo comments to this file."
+                raise TypeError(reason)
             comment = parser.to_string().encode(errors="replace")
-            append_archive.comment = comment  # type: ignore
+            append_archive.comment = comment
 
     def close(self):
         """Close the open archive."""
         try:
             if self._archive and hasattr(self._archive, "close"):
-                self._archive.close()  # type: ignore
+                self._archive.close()
         except Exception as exc:
             LOG.warning(f"closing archive: {exc}")
         finally:
@@ -324,32 +375,34 @@ class ComicArchive:
         return self._metadata.get_num_pages()
 
     @_archive_close
-    def get_pages(self, page_from):
+    def get_pages(self, page_from=0, page_to=-1, to_pixmap=False):
         """Generate all pages starting with page number."""
         self._ensure_page_metadata()
-        pagenames = self._metadata.get_pagenames_from(page_from)
+        pagenames = self._metadata.get_pagenames_from(page_from, page_to)
         if pagenames:
             for pagename in pagenames:
-                yield self._archive_readfile(pagename)
+                yield self._archive_readfile(pagename, to_pixmap)
 
     @_archive_close
-    def get_page_by_filename(self, filename):
+    def get_page_by_filename(self, filename, to_pixmap=False):
         """Return data for a single page by filename."""
-        return self._archive_readfile(filename)
+        return self._archive_readfile(filename, to_pixmap)
 
     @_archive_close
-    def get_page_by_index(self, index):
+    def get_page_by_index(self, index, to_pixmap=False):
         """Get the page data by index."""
         self._ensure_page_metadata()
         filename = self._metadata.get_pagename(index)
-        return self._archive_readfile(filename)
+        return self._archive_readfile(filename, to_pixmap)
 
-    def _extract_page(self, path, fn):
+    def _extract_page(self, path, fn, to_pixmap=False):
+        if self._archive_cls == PDFFile:
+            path = path.with_suffix(PDFFile.SUFFIX)
         with path.open("wb") as page_file:
-            page_file.write(self._archive_readfile(fn))
+            page_file.write(self._archive_readfile(fn, to_pixmap))
 
     @_archive_close
-    def extract_pages(self, page_from, root_path="."):
+    def extract_pages(self, page_from=None, page_to=None, root_path="."):
         """Extract pages from archive and write to a path."""
         root_path = Path(root_path)
         if not root_path.is_dir():
@@ -359,7 +412,7 @@ class ComicArchive:
             )
             raise ValueError(reason)
         self._ensure_page_metadata()
-        pagenames = self._metadata.get_pagenames_from(page_from)
+        pagenames = self._metadata.get_pagenames_from(page_from, page_to)
         if pagenames:
             for fn in pagenames:
                 if self._config.dry_run:
@@ -394,7 +447,7 @@ class ComicArchive:
             return None
         data = None
         try:
-            data = self._archive_readfile(cover_fn)
+            data = self._archive_readfile(cover_fn, True)
         except Exception as exc:
             LOG.warning(f"{self._path} reading cover: {cover_fn}: {exc}")
         return data
@@ -411,9 +464,11 @@ class ComicArchive:
     def _get_comment(self):
         """Get the comment from the archive."""
         comment = b""
-        if self._archive_cls != tarfile_open:
+        if self._archive_cls in self._COMMENT_ARCHIVE_TYPES:
             if not self._config.delete_tags:
-                comment = self._get_archive().comment  # type: ignore
+                archive = self._get_archive()
+                if isinstance(archive, self._COMMENT_ARCHIVE_TYPES):
+                    comment = archive.comment
             if isinstance(comment, str):
                 comment = comment.encode(errors="replace")
 
@@ -434,7 +489,7 @@ class ComicArchive:
         compress = ZIP_DEFLATED if IMAGE_EXT_RE.search(fn) is None else ZIP_STORED
         zf.writestr(
             fn,
-            self._archive_readfile(fn),
+            self._archive_readfile(fn, True),
             compress_type=compress,
             compresslevel=9,
         )
@@ -481,6 +536,17 @@ class ComicArchive:
                 old_path.unlink()
                 LOG.info(f"removed: {old_path}")
 
+    def _write_pdf_metadata(self, parser):
+        """Write PDF Metadata."""
+        if self._config.dry_run:
+            LOG.info("Not writing PDF metadata.")
+            return
+        if self._archive_cls != PDFFile:
+            return
+        pdf_md = parser.to_dict()
+        with PDFFile(self._path) as archive:
+            archive.save_metadata(pdf_md)
+
     def write_metadata(self, md_class, metadata=None, recompute_page_sizes=True):
         """Write metadata using the supplied parser class."""
         if self._config.dry_run:
@@ -488,13 +554,15 @@ class ComicArchive:
             return
         if metadata is None:
             metadata = self.get_metadata()
-        parser = md_class(metadata=metadata)
+        parser = md_class(path=self._path, metadata=metadata)
         if recompute_page_sizes and isinstance(parser, ComicInfoXml):
             self.compute_pages_tags()
         if isinstance(parser, (ComicXml, CoMet)):
             self.recompress(parser.FILENAME, parser.to_string())
         elif isinstance(parser, ComicBookInfo):
             self._write_cbi_comment(parser)
+        elif isinstance(parser, PDFParser):
+            self._write_pdf_metadata(parser)
         else:
             reason = f"Unsupported metadata writer {md_class}"
             raise TypeError(reason)
@@ -502,10 +570,12 @@ class ComicArchive:
 
     def _get_write_parsers(self):
         """Convert config.write list into parser class list."""
+        is_pdf = self._archive_cls == PDFFile
         md_class_dict = {
-            ComicInfoXml: self._config.write_comicinfoxml,
-            ComicBookInfo: self._config.write_comicbookinfo,
-            CoMet: self._config.write_comet,
+            ComicInfoXml: self._config.write_comicinfoxml and not is_pdf,
+            ComicBookInfo: self._config.write_comicbookinfo and not is_pdf,
+            CoMet: self._config.write_comet and not is_pdf,
+            PDFParser: self._config.write_pdf and is_pdf,
         }
 
         md_class_list = []
@@ -562,7 +632,7 @@ class ComicArchive:
         """Recompute the tag image sizes for ComicRack."""
         infolist, _ = self._archive_infolist()
         metadata = self.get_metadata()
-        parser = ComicInfoXml(metadata=metadata, path=self._path)
+        parser = ComicInfoXml(path=self._path, metadata=metadata)
         parser.compute_pages_tags(infolist)
         self._metadata.metadata["pages"] = parser.metadata.get("pages")
 
@@ -577,9 +647,9 @@ class ComicArchive:
         """Print raw metadtata."""
         self._set_raw_metadata()
         for key, val in self._raw.items():
-            print("-" * 10, key, "-" * 10)  # noqa T201
+            print("-" * 10, key, "-" * 10)  # noqa: T201
             print_val = val.decode(errors="replace") if isinstance(val, bytes) else val
-            print(print_val)  # noqa T201
+            pprint(print_val)  # noqa: T203
 
     def get_path(self):
         """Get the path for the archive."""