+++ comicbox/metadata/filename.py
@@ -203,7 +203,7 @@ class FilenameMetadata(ComicBaseMetadata):
             tokens.append(token)
         return " ".join(tokens)
 
-    def to_file(self, path, dry_run=False):
+    def to_file(self, path, dry_run=True):
         """Rename this file according to our favorite naming scheme."""
         name = self.to_string()
         new_path = path.parent / Path(name + path.suffix)