+++ tests/test_filename_parse.py
@@ -11,7 +11,7 @@ from tests.test_filenames import FNS
 def test_parse_filename(item):
     """Test filename parsing."""
     fn, defined_fields = item
-    parser = FilenameMetadata(path=fn)
+    parser = FilenameMetadata(path=fn, metadata_path=fn)
     res = parser.metadata
     matched_fields = set()
     unmatched_fields = []